Hashing done
I feel dirty doing c casting


Former-commit-id: 1c33327439adbe126ee8a077d726dcdb9e3e5844

+++ database/include/database.h
@@ -8,7 +8,7 @@
 #include <vector>
   
 namespace db {
-using UrlId = int;
+using UrlId = uint64_t;
 
 class InvIndexStorage {
 public:
